-pc_jobid2mapid and pc_mapid2jobid now return int, and will return -1 when the class couldn't be deciphered. - pc_authok will change you to notice if your current class is invalid. - rewrote pc_jobchange to be much more simple and less of a hassle by using pc_jobid2mapid and pc_mapid2jobid as necessary.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@5520 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/pc.h
@@ -198,8 +198,8 @@ struct pc_base_job{
 
 struct pc_base_job pc_calc_base_job(int b_class);//ì]ê∂Ç‚ó{éqêEÇÃå≥ÇÃêEã∆Çï‘Ç∑
 int pc_calc_base_job2(int b_class);	// Celest
-unsigned short pc_jobid2mapid(unsigned short b_class);	// Skotlex
-unsigned short pc_mapid2jobid(unsigned short class_, int sex);	// Skotlex
+int pc_jobid2mapid(unsigned short b_class);	// Skotlex
+int pc_mapid2jobid(unsigned short class_, int sex);	// Skotlex
 
 char * job_name(int class_);
